Refactor form elements and form layout including controller
