+++ example-scss/_nx-input-radio.scss
@@ -24,8 +24,8 @@ input.nx-radio__input {
   display: none;
 
   &:checked + .nx-radio__circle::after {
-    background-color: $iq-slate;
-    border: 1px solid $iq-slate;
+    background-color: $nx-slate;
+    border: 1px solid $nx-slate;
     border-radius: 50%;
     -webkit-transform: scale(.5);
     -moz-transform: scale(.5);
@@ -37,7 +37,7 @@ input.nx-radio__input {
 }
 
 .nx-radio__circle {
-  color: $iq-off-black;
+  color: $nx-off-black;
   cursor: pointer;
   display: inline-block;
 }
@@ -54,24 +54,24 @@ input.nx-radio__input {
 }
 
 .nx-radio__circle::before {
-  border: 1px solid $iq-slate;
+  border: 1px solid $nx-slate;
   border-radius: 50%;
 }
 
 .nx-radio--disabled {
-  color: $iq-grey-4;
+  color: $nx-grey-4;
 
   .nx-radio__circle {
-    color: $iq-slate;
+    color: $nx-slate;
     cursor: default;
   
     &::before {
-      border-color: $iq-grey-4;
+      border-color: $nx-grey-4;
     }
   }
 
   .nx-radio__input:checked + .nx-radio__circle::after {
-    background-color: $iq-grey-4;
-    border-color: $iq-grey-4;
+    background-color: $nx-grey-4;
+    border-color: $nx-grey-4;
   }
 }
